Subtracting 773 meters from final distance in circle measurement. This seemed to be roughly the difference when creating an alert directly on my address. I believe we would need to be using a more precise number than rounding off or even more than 6-8 decimal places to make this work. I could be wrong.
So using the "new" way to get addresses in alerts, I have the texting and email setup for send new alerts and updates.
I also implemented the cancel button on the update alert page.
I also got the table on the update alert page working. On update, it fills the form with the most recent update info not including the start time of the original alert. Then updates the table adding the new update to the table.
Table also only displays if the table has data. Will not show empty table anymore.

I have the texting section commented out in the create and update alerts just to avoid spending unncessary money on texts.